split up layers and materials
