+++ EasyReflectometry/sample/__init__.py
@@ -2,12 +2,12 @@ from .assemblies.gradient_layer import GradientLayer
 from .assemblies.multilayer import MultiLayer
 from .assemblies.repeating_multilayer import RepeatingMultiLayer
 from .assemblies.surfactant_layer import SurfactantLayer
-from .elementals.layer import Layer
-from .elementals.layer import LayerApm
 from .elementals.layer_collection import LayerCollection
-from .elementals.material import Material
-from .elementals.material import MaterialMixture
+from .elementals.layers.layer import Layer
+from .elementals.layers.layer_apm import LayerApm
 from .elementals.material_collection import MaterialCollection
+from .elementals.materials.material import Material
+from .elementals.materials.material_mixture import MaterialMixture
 from .sample import Sample
 
 __all__ = (
