+++ EasyReflectometry/sample/assemblies/gradient_layer.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 from easyCore.Fitting.Constraints import ObjConstraint
 from numpy import arange
 
-from ..elementals.layer import Layer
-from ..elementals.material import Material
+from ..elementals.layers.layer import Layer
+from ..elementals.materials.material import Material
 from .multilayer import MultiLayer
