+++ EasyReflectometry/sample/elementals/layer_collection.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 __author__ = 'github.com/arm61'
 
 from .base import BaseElementCollection
-from .layer import Layer
+from .layers.layer import Layer
 
 
 class LayerCollection(BaseElementCollection):
