+++ EasyReflectometry/sample/elementals/material_collection.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 __author__ = 'github.com/arm61'
 
 from .base import BaseElementCollection
-from .material import Material
-from .material import MaterialMixture
+from .materials.material import Material
+from .materials.material_mixture import MaterialMixture
 
 
 class MaterialCollection(BaseElementCollection):
