+++ tests/sample/elementals/test_layer_collection.py
@@ -9,9 +9,9 @@ import unittest
 from numpy.testing import assert_equal
 
 from EasyReflectometry.sample.assemblies.repeating_multilayer import RepeatingMultiLayer
-from EasyReflectometry.sample.elementals.layer import Layer
+from EasyReflectometry.sample.elementals.layers.layer import Layer
 from EasyReflectometry.sample.elementals.layer_collection import LayerCollection
-from EasyReflectometry.sample.elementals.material import Material
+from EasyReflectometry.sample.elementals.materials.material import Material
 
 
 class TestLayerCollection(unittest.TestCase):
